update content routes and fragments
Update default content page fragment to not use factor and just set BCL to false.  If we want it to true can hard code it or re introduct the env. Also setup a with leades info version with a factory that takes a leaders alias in order to get the compnies isLeader flag